requirements.txt
Django~=5.2.1
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
psycopg2-binary==2.9.9

djangorestframework~=3.16.0
rest_framework_simplejwt